Move box-plot to its own component
